Split components into individual modules
