+++ poucave/html/index.html
@@ -22,6 +22,6 @@
   <script type="text/javascript" src="js/marked.min.js"></script>
   <script type="text/javascript" src="js/timeago.min.js"></script>
   <script type="text/javascript" src="js/plotly.min.js"></script>
-  <script type="module" src="js/app.js"></script>
+  <script type="module" src="js/app/index.mjs"></script>
 </body>
 </html>
