+++ poucave/html/js/app/contexts/FocusedCheck.mjs
@@ -1,0 +1,9 @@
+import { createContext } from "../../htm_preact.mjs";
+
+const FocusedCheck = createContext({
+  project: null,
+  name: null,
+  setValue: () => {},
+});
+
+export default FocusedCheck;
